Interfaces relocation
