+++ src/interfaces/SearchInterface.ts
